recruiter_ai_service/requirements.txt
# Recruiter AI Service Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AI/ML
openai>=1.6.1
numpy==1.24.3

# LlamaIndex (Document-centric AI framework) - v0.12+ with modular structure
llama-index-core>=0.12.0,<0.13.0
llama-index-vector-stores-supabase==0.3.0
llama-index-embeddings-openai>=0.3.0
llama-index-llms-openai>=0.4.0
llama-index>=0.12.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Vector Databases
pgvector==0.2.4
pinecone-client==2.2.4

# Queue and Caching
redis[hiredis]==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1

# Logging and Monitoring
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0